requirements.txt
streamlit>=1.22
groq>=0.6.0
python-dotenv>=1.0
pymupdf>=1.26.0    # fitz
pdfplumber>=0.7.6
pillow>=11.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pandas>=2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-magic>=0.4.27
